requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.8
django-cors-headers==3.2.1
djangorestframework==3.11.0
drf-yasg==1.17.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
packaging==20.0
pathspec==0.7.0
Pillow==7.0.0
psycopg2-binary==2.8.4
pyparsing==2.4.6
pytz==2019.3
regex==2019.12.20
requests==2.22.0
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.2.0
six==1.13.0
sqlparse==0.3.0
toml==0.10.0
transliterate==1.10.2
typed-ast==1.4.0
uritemplate==3.0.1
urllib3==1.25.7
